4.2.1 (#81)
* Removed unnecessary `bot.addAutowiringObject` method.
* Enhanced `userData`/`chatData` get method with generic wrapped with catching for safety.
* Added `allowedUpdates` for long-polling since it necessary for getting special type updates.
* Fixed `giveawayMessageId` type: `Int` > `Long`.
* Added `replyParameters(messageId, block: ReplyParameters.() -> Unit)` fun-shortcut for changing replyParameters.
* Optimized action internal type inference.

+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/types/GiveawayWinners.kt
@@ -4,7 +4,7 @@ import eu.vendeli.tgbot.types.chat.Chat
 
 data class GiveawayWinners(
     val chat: Chat,
-    val giveawayMessageId: Int,
+    val giveawayMessageId: Long,
     val winnersSelectionDate: Long,
     val winnerCount: Int,
     val winners: List<User>,
